Encounters schema for NNDD. Fixes #476.
