+++ app/models/encounters_schema.rb
@@ -1,0 +1,7 @@
+class EncountersSchema < EntitySchema
+
+  def scopes
+    Cdx::Fields.encounter.core_field_scopes
+  end
+
+end
