+++ config/routes.rb
@@ -109,6 +109,11 @@ Rails.application.routes.draw do
       end
     end
     match 'encounters(.:format)' => "encounters#index", via: [:get, :post]
+    resources :encounters, only: [] do
+      collection do
+        get :schema
+      end
+    end
     resources :devices, only: [] do
       resources :messages, only: [:create ], shallow: true
       match 'tests' => "messages#create", via: :post # For backwards compatibility with Qiagen-Esequant-LR3
