+++ spec/models/encounters_schema_spec.rb
@@ -1,0 +1,13 @@
+require 'spec_helper'
+
+describe EncountersSchema do
+
+  context "encounters fields" do
+
+    it "renders all encounters scopes" do
+      expect(EncountersSchema.new.build['properties'].keys).to contain_exactly("institution", "patient", "encounter")
+    end
+
+  end
+
+end
